chore(PR Review):  migrate payload items to functions instead of classes
