+++ packages/web-sdk/src/transports/otlp/transform/index.ts
@@ -1,12 +1,12 @@
-export { attributeValueType, toAttribute, toNestedAttributes } from './attributeUtils';
+export { AttributeValueType, toAttribute, toNestedAttributes } from './attributeUtils';
 export { SemanticBrowserAttributes, faroResourceAttributes } from './semanticResourceAttributes';
-export { LogRecordFactory } from './LogRecordFactory';
-export { LogRecord } from './LogLogRecord';
-export { ErrorLogRecord as ErrorRecord } from './ErrorLogRecord';
+export { LogRecordFactory } from './deprecated/LogRecordFactory';
+export { LogRecord } from './deprecated/LogLogRecord';
+export { ErrorLogRecord as ErrorRecord } from './deprecated/ErrorLogRecord';
 export { EventRecord } from './EventLogRecord';
-export { Payload } from './Payload';
-export { Resource } from './Resource';
-export { ResourceLog } from './ResourceLog';
-export { Scope } from './Scope';
-export { ScopeLog } from './ScopeLog';
+export { Payload } from './deprecated/Payload';
+export { Resource } from './deprecated/Resource';
+export { ResourceLog } from './deprecated/ResourceLog';
+export { Scope } from './deprecated/Scope';
+export { ScopeLog } from './deprecated/ScopeLog';
 export { Attribute, AttributeTypes, FaroResourceAttributes, PayloadMember, ResourceLogs } from './types';
